requirements.txt
gspread
google-oauth2-tool
python-dotenv